Probabilistic Models

BayesianExperiments.ProbabilisticModel — Type

ProbabilisticModel

ProbabilisticModel is a a model of the parameters that we are interested in. The model is defined by its prior distribution and likelihood function.

BayesianExperiments.ConjugateModel — Type

ConjugateModel <: ProbabilisticModel

ConjugateModel is a ProbabilisticModel with a conjugate prior of the corresponding likelihood function.

BayesianExperiments.BernoulliModel — Type

BernoulliModel <: ConjugateModel

Bernoulli likelihood with Beta distribution as the conjugate prior.

BernoulliModel(α, β)              # construct a BernoulliModel

update!(model, stats)             # update model with statistics from data
sample_post(model, numsamples)    # sampling from the posterior distribution
sample_stats(model, numsamples)   # sampling statistics from the data generating distribution

BayesianExperiments.ExponentialModel — Type

ExponentialModel <: ConjugateModel

Exponential likelihood with Gamma distribution as the conjugate prior.

ExponentialModel(α, β)            # construct a ExponentialModel

update!(model, stats)             # update model with statistics from data
sample_post(model, numsamples)    # sampling from the posterior distribution
sample_stats(model, numsamples)   # sampling statistics from the data generating distribution

BayesianExperiments.NormalModel — Type

NormalModel <: ConjugateModel

Normal likelihood and Normal Inverse Gamma distribution as the conjugate prior.

Parameters

μ: mean of normal distribution

v: scale variance of Normal

α: shape of Gamma distribution

θ: scale of Gamma distribution

NormalModel(μ, v, α, θ)           # construct a NormalModel

update!(model, stats)             # update model with statistics from data
sample_post(model, numsamples)    # sampling from the posterior distribution
sample_stats(model, numsamples)   # sampling statistics from the data generating distribution

References

The update rule for Normal distribution is based on this lecture notes.

BayesianExperiments.LogNormalModel — Type

LogNormalModel(μ, v, α, θ)

A model with Normal likelihood and Normal Inverse distribution with log transformed data. Notice LogNormal in Distributions.jl takes mean and standard deviation of $\log(x)$ instead of $x$ as the input parameters.

LogNormalModel(μ, v, α, θ) # construct a LogNormalModel

tolognormalparams(μ_logx, σ²_logx) # convert normal parameters to log-normal parameters
update!(model, stats)              # update model with statistics from data
sample_post(model, numsamples)     # sampling from the posterior distribution
sample_stats(model, numsamples)    # sampling statistics from the data generating distributio

BayesianExperiments.ChainedModel — Type

ChainedModel <: ProbabilisticModel

ChainedModel is a combination of ConjugateModels chained by the specified operator. It can be used to model a multiple step process.

Stopping Rules

BayesianExperiments.ExpectedLossThresh — Type

ExpectedLossThresh <: StoppingRule

The experiment has a winning model if the model has the smallest posterior expected loss, and its expected loss value is below the threshold.

References

Definition of the expected loss on Wikipedia

BayesianExperiments.ProbabilityBeatAllThresh — Type

ProbabilityBeatAllThresh <: StoppingRule

The experiment has a winning model if probability of that model's posterior samples is larger than the alternative models is above the threshold.

Experiment

BayesianExperiments.Experiment — Type

Experiment

An experiment has the models to compare, a stopping rule to make decision.

BayesianExperiments.ExperimentABN — Type

ExperimentABN{T,n} <: Experiment

An experiment with stopping rule of type T and n models. Models must have the same ProbabilisticModel type.

BayesianExperiments.ExperimentAB — Function

ExperimentAB(models, rule; modelnames=nothing) where T <: StoppingRule

An experiment of ExperimentABN with two models.

Simulation

BayesianExperiments.Simulation — Type

Simulation(experiment, parameters, datagendists, maxsteps, onestepsizes, minsteps)

A simulation setup includes the experiment, data generating distributions, max number of steps and minimum number of steps.